ENH: rename cell_classes to cells_default.py
