+++ hnn_core/__init__.py
@@ -3,7 +3,7 @@ from .drives import drive_event_times
 from .params import Params, read_params
 from .network import Network
 from .cell_response import CellResponse, read_spikes
-from .cell_classes import pyramidal, basket
+from .cells_default import pyramidal, basket
 from .parallel_backends import MPIBackend, JoblibBackend
 
 __version__ = '0.2.dev0'
